+++ tasks/noci_test.sh
@@ -5,7 +5,7 @@ EXIT_STATE=0
 # tests that aren't run on CI
 
 # jasmine specs with @noCI tag
-npm run citest-jasmine --  tests/*_test.js --tags noCI || EXIT_STATE=$?
+npm run test-jasmine -- --tags=noCI --nowatch || EXIT_STATE=$?
 
 # mapbox image tests take too much resources on CI
 npm run test-image -- mapbox_* || EXIT_STATE=$?