+++ test/jasmine/assets/mouse_event.js
@@ -1,3 +1,5 @@
+var Lib = require('@src/lib');
+
 module.exports = function(type, x, y, opts) {
     var fullOpts = {
         bubbles: true,
@@ -9,12 +11,24 @@ module.exports = function(type, x, y, opts) {
     if(opts && opts.buttons) {
         fullOpts.buttons = opts.buttons;
     }
+    if(opts && opts.altKey) {
+        fullOpts.altKey = opts.altKey;
+    }
+    if(opts && opts.ctrlKey) {
+        fullOpts.ctrlKey = opts.ctrlKey;
+    }
+    if(opts && opts.metaKey) {
+        fullOpts.metaKey = opts.metaKey;
+    }
+    if(opts && opts.shiftKey) {
+        fullOpts.shiftKey = opts.shiftKey;
+    }
 
     var el = (opts && opts.element) || document.elementFromPoint(x, y),
         ev;
 
     if(type === 'scroll') {
-        ev = new window.WheelEvent('wheel', opts);
+        ev = new window.WheelEvent('wheel', Lib.extendFlat({}, fullOpts, opts));
     } else {
         ev = new window.MouseEvent(type, fullOpts);
     }